+++ src/cmd/cli/command/estimate_test.go
@@ -148,7 +148,7 @@ $32.85   730 Hours         shared   AmazonEC2 USW2-NatGateway-Hours
 Estimated Monthly Cost: $118.55 (+ usage)
 
 Estimate does not include taxes or Discount Programs.
-To estimate other modes, use defang estimate --mode=affordable|balanced|high_availability
+To estimate other modes, use defang estimate --mode=AFFORDABLE|BALANCED|HIGH_AVAILABILITY
 `
 
 	outputLines := strings.Split(term.StripAnsi(stdout.String()), "\n")
